requirements.txt
acres==0.2.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==24.3.0
beautifulsoup4==4.13.3
bs4==0.0.2
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.8
cohere==5.13.8
colorama==0.4.6
comm==0.2.2
comtypes==1.4.10
configobj==5.0.9
configparser==7.1.0
cryptography==44.0.0
debugpy==1.8.11
decorator==5.1.1
distro==1.9.0
etelemetry==0.3.1
exceptiongroup==1.2.2
executing==2.1.0
fastavro==1.10.0
filelock==3.16.1
frontend==0.0.3
frozenlist==1.5.0
fsspec==2024.12.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
greenlet==3.1.1
gTTS==2.5.4
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.18.1
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.2
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.13
langchain-core==0.3.28
langchain-openai==0.2.14
langchain-text-splitters==0.3.4
langgraph==0.2.69
langgraph-checkpoint==2.0.10
langgraph-sdk==0.1.51
langsmith==0.2.6
looseversion==1.3.0
lxml==5.3.0
matplotlib-inline==0.1.7
mmh3==4.1.0
msgpack==1.1.0
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.2.1
nibabel==5.3.2
nipype==1.9.2
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.58.1
orjson==3.10.12
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
parso==0.8.4
pathlib==1.0.1
pdfminer==20191125
pdfminer.six==20240706
pillow==11.0.0
pinecone==5.4.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
pinecone-text==0.9.0
platformdirs==4.3.6
polars==1.17.1
prompt_toolkit==3.0.48
propcache==0.2.1
prov==2.0.1
psutil==6.1.1
pure_eval==0.2.3
puremagic==1.28
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.4
pydantic_core==2.27.2
pydot==3.0.3
pygame==2.6.1
Pygments==2.18.0
PyMuPDF==1.25.1
pyparsing==3.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-pptx==1.0.2
pyttsx3==2.98
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.2
pyzmq==26.2.0
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rsa==4.9
scikit-learn==1.6.0
scipy==1.13.1
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.14.1
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.42.0
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traits==6.4.3
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
wget==3.2
XlsxWriter==3.2.0
yarl==1.18.3
zipp==3.21.0
openai-whisper==20240930
thefuzz==0.22.1
pdf2image==1.17.0
deep-translator==1.11.4